Move storage field to ty module.
