+++ sway-core/src/control_flow_analysis/flow_graph/namespace.rs
@@ -1,8 +1,9 @@
 use super::{EntryPoint, ExitPoint};
-use crate::language::CallPath;
-use crate::semantic_analysis::declaration::TyStorageField;
-use crate::type_system::TypeInfo;
-use crate::Ident;
+use crate::{
+    language::{ty, CallPath},
+    type_system::TypeInfo,
+    Ident,
+};
 use petgraph::prelude::NodeIndex;
 use std::collections::HashMap;
 
@@ -122,7 +123,7 @@ impl ControlFlowNamespace {
         }
     }
 
-    pub(crate) fn insert_storage(&mut self, field_nodes: Vec<(TyStorageField, NodeIndex)>) {
+    pub(crate) fn insert_storage(&mut self, field_nodes: Vec<(ty::TyStorageField, NodeIndex)>) {
         for (field, node) in field_nodes {
             self.storage.insert(field.name, node);
         }
