+++ sway-core/src/semantic_analysis/ast_node/declaration/storage.rs
@@ -8,7 +8,7 @@ use crate::{
     semantic_analysis::{
         TyStructField, TypeCheckedStorageAccess, TypeCheckedStorageAccessDescriptor,
     },
-    type_system::{look_up_type_id, TypeId, TypeInfo},
+    type_system::{TypeId, TypeInfo},
     AttributesMap, Ident,
 };
 use fuel_tx::StorageSlot;
@@ -17,7 +17,7 @@ use sway_ir::{Context, Module};
 use sway_types::{state::StateIndex, Span, Spanned};
 
 impl ty::TyStorageDeclaration {
-    pub fn new(fields: Vec<TyStorageField>, span: Span, attributes: AttributesMap) -> Self {
+    pub fn new(fields: Vec<ty::TyStorageField>, span: Span, attributes: AttributesMap) -> Self {
         ty::TyStorageDeclaration {
             fields,
             span,
@@ -29,7 +29,7 @@ impl ty::TyStorageDeclaration {
     pub fn apply_storage_load(
         &self,
         fields: Vec<Ident>,
-        storage_fields: &[TyStorageField],
+        storage_fields: &[ty::TyStorageField],
     ) -> CompileResult<(TypeCheckedStorageAccess, TypeId)> {
         let mut errors = vec![];
         let warnings = vec![];
@@ -41,11 +41,11 @@ impl ty::TyStorageDeclaration {
         let (ix, initial_field_type) = match storage_fields
             .iter()
             .enumerate()
-            .find(|(_, TyStorageField { name, .. })| name == &first_field)
+            .find(|(_, ty::TyStorageField { name, .. })| name == &first_field)
         {
             Some((
                 ix,
-                TyStorageField {
+                ty::TyStorageField {
                     type_id: r#type, ..
                 },
             )) => (StateIndex::new(ix), r#type),
@@ -123,7 +123,7 @@ impl ty::TyStorageDeclaration {
         self.fields
             .iter()
             .map(
-                |TyStorageField {
+                |ty::TyStorageField {
                      ref name,
                      type_id: ref r#type,
                      ref span,
@@ -167,28 +167,7 @@ impl ty::TyStorageDeclaration {
     }
 }
 
-#[derive(Clone, Debug, Eq)]
-pub struct TyStorageField {
-    pub name: Ident,
-    pub type_id: TypeId,
-    pub type_span: Span,
-    pub initializer: ty::TyExpression,
-    pub(crate) span: Span,
-    pub attributes: AttributesMap,
-}
-
-// NOTE: Hash and PartialEq must uphold the invariant:
-// k1 == k2 -> hash(k1) == hash(k2)
-// https://doc.rust-lang.org/std/collections/struct.HashMap.html
-impl PartialEq for TyStorageField {
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-            && look_up_type_id(self.type_id) == look_up_type_id(other.type_id)
-            && self.initializer == other.initializer
-    }
-}
-
-impl TyStorageField {
+impl ty::TyStorageField {
     pub(crate) fn get_initialized_storage_slots(
         &self,
         context: &mut Context,
